pyproject.toml
[project]
name = "app"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "django>=4.2,<5.0",
    "djangorestframework>=3.14",
    "djangorestframework-simplejwt>=5.3",
    "django-cors-headers>=4.3",
    "django-filter>=23.5",
    "whitenoise>=6.6",
    "pydantic>=2.0",
    "claude-agent-sdk>=0.1.25",
    "dj-database-url>=2.0",
    "psycopg2-binary>=2.9,<3.0",
    "gunicorn>=21.2.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.0",
    "pytest-django>=4.8",
    # Linting and formatting
    "black>=24.0",
    "isort>=5.13",
    "prospector>=1.10",
    "pylint-django>=2.5",
    "mypy>=1.8",
    "django-stubs>=4.2",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
pythonpath = ["."]
testpaths = ["."]
norecursedirs = [".*", "node_modules", ".venv"]
addopts = "--rootdir=."
